+++ R/garcon.R
@@ -109,15 +109,6 @@ Garcon <- R6::R6Class(
       cat("A garcon\n")
 		},
 #' @details
-#' Stop the garçon.
-#' 
-#' @examples
-#' \dontrun{Garcon$new("img")$set(30)$stop()}
-    stop = function() {
-      .Deprecated("close", package = "waiter", "`stop` is deprecated in favour of `close`")
-      private$.session$sendCustomMessage("garcon-stop", list(image = private$.image))
-    },
-#' @details
 #' Close the garçon.
 #' 
 #' @examples